build: rename error stack trace → stack trace
