+++ packages/goto/src/evasions/stack-traces.js
